Added tags to immersives
- Refactored tags component to use defaults pattern
- Updated usages of tags component
- Added `ImmersiveTags` component
- Updated tests and stories to use main item fixture
