+++ apps-rendering/src/components/Layout/InteractiveImmersiveLayout.tsx
@@ -90,7 +90,7 @@ const InteractiveImmersiveLayout: FC<Props> = ({ item, children }) => {
 				<Body format={item}>{children}</Body>
 				{epicContainer}
 				<section className="js-tags" css={articleWidthStyles}>
-					<Tags tags={item.tags} format={item} />
+					<Tags item={item} />
 				</section>
 			</article>
 			<section css={onwardStyles}>
