+++ apps-rendering/src/components/Layout/LiveLayout.tsx
@@ -192,7 +192,7 @@ const LiveLayout: FC<Props> = ({ item }) => {
 				</GridItem>
 			</main>
 			<section css={articleWidthStyles}>
-				<Tags tags={item.tags} format={item} />
+				<Tags item={item} />
 			</section>
 			<section css={onwardStyles}>
 				<RelatedContent content={item.relatedContent} />
