+++ apps-rendering/src/components/Layout/StandardLayout.tsx
@@ -142,7 +142,7 @@ const StandardLayout: FC<Props> = ({ item, children }) => {
 				</Body>
 				{epicContainer}
 				<section className="js-tags" css={articleWidthStyles}>
-					<Tags tags={item.tags} format={item} />
+					<Tags item={item} />
 				</section>
 			</article>
 			<section css={onwardStyles}>
